move about to its own directory
